Move AutocompleteField and revert changesMove AutocompleteField and revert changes
